refactor: split types.go
